BasicScoreboard: Allow the comparator class to be extended so we can send data to a coverage collector.
You can now extend the comparator overriding the send_data_out_on_pass
method. This method takes both the predicted transaction and the actual
transaction. The child class can then define an analysis port and write
one of those transactions or in fact whatever data it wants to the
analysis port.

This allows us to have a coverage collector that receives data after a
successfull transaction.